requirements.txt
orb==0.9.4
psycopg2-binary==2.9.3
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.1
